Stream.unwrap: move to prelude and combinators
