+++ kyo-core/shared/src/main/scala/kyo/StreamCoreExtensions.scala
@@ -566,13 +566,6 @@ object StreamCoreExtensions:
 
     end extension
 
-    extension [V, S, S2](stream: Stream[V, S] < S2)
-        /** Takes a Stream[V, S] in the context of S2 (i.e. Stream[V, S] < S2) and returns a Stream that fuses together both effect contexts
-          * S and S2 into a single Stream[V, S & S2].
-          */
-        inline def unwrap: Stream[V, S & S2] = Stream(stream.map(_.emit))
-    end extension
-
 end StreamCoreExtensions
 
 export StreamCoreExtensions.*
